Changed logic abit, and package name
