--- src/com/company/Main.java
@@ -1,9 +1,0 @@
-package com.company;
-
-public class Main {
-
-    public static void main(String[] args) throws InterruptedException {
-        SpotSaver ss = new SpotSaver();
-        ss.Execute();
-    }
-}
